storage/disk: impose same locking regime used with inmem
With this setup, we'll ensure:

- there is only one open write txn at a time
- there are any number of open read txns at a time
- writes are blocked when reads are inflight
- during a commit (and triggers being run), no read txns can be created

This is to ensure the same atomic policy update semantics when using
'disk" as we have with "inmem". We're basically opting out of badger's
currency control and transactionality guarantees. This is because we
cannot piggy back on that to ensure the atomic update we want.

There might be other ways -- using subscribers, and blocking in some
other place -- but this one seems preferrable since it mirrors inmem.

Part of the problem is ErrTxnTooLarge, and committing and renewing
txns when it occurs: that, which is the prescribed solution to txns
growing too big, also means that reads can see half of the "logical"
transaction having been committed, while the rest is still getting
processed.

Another approach would have been using `WriteBatch`, but that won't
let us read from the batch, only apply Set and Delete operations.
We currently need to read (via an iterator) to figure out if we need
to delete keys to replace something in the store:

For example, if we have

    data:
      foo:
        baz: 100

that'll be one key: `/data/foo/baz`, with value 100. If we then
store value 200 at `data.foo`, we'll iterate the `/data/foo` prefix,
remove any key we find (`/data/foo/baz`), and store the new one.

There is no DropPrefix operation on the badger txn, or the WriteBatch
API.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>