topdown/copyprop: deduplicate and check expressions when adding to result body
Before, we'd have added expressions multiple times. Also, it was possible to
generate rule bodies that would never work out because they carried both
expr and not-expr for the same expr.

Now, we're checking these two things before adding an expression to the result
in copypropagation.

Fixes #4516 for simple cases. It's still the case that we may end up with
impossible bodies, like

    not input.foo
    input.foo = "bar"

but at least we won't have

    input.foo.bar
    not input.foo.bar

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>